refactor: add scss styling partial to every component
