+++ src/Assets/Components/BottomNav/_BottomNav.scss
@@ -1,0 +1,40 @@
+.mobile-nav-items {
+  display: block;
+  @include md {
+    display: none;
+  }
+
+  &-container {
+    display: flex;
+    justify-content: space-around;
+    position: fixed;
+    bottom: 0px;
+    width: 100%;
+    padding: 0.313rem 0.5rem 0rem 0.5rem;
+    border-top: 0.063rem solid $colors-dark-cardBorder;
+    background: $colors-dark-backgroundAlt;
+    z-index: 20;
+
+    .menu-item {
+      button {
+        @include aAsButton;
+        display: flex;
+        flex-flow: column;
+        text-decoration: none;
+        align-items: center;
+        justify-content: center;
+        color: $colors-dark-textSubtle;
+        font-size: 0.625rem;
+        height: 2.75rem;
+        gap: 0.125rem;
+
+        img {
+          // height: 1.375rem;
+          // width: 1.313rem;
+          height: 1.313rem;
+          width: 1.25rem;
+        }
+      }
+    }
+  }
+}
