+++ src/Assets/Components/Header/NavBar/_NavBar.scss
@@ -1,0 +1,208 @@
+.nav-container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  height: 3.5rem;
+  background-color: $colors-dark-backgroundAlt;
+  color: $colors-dark-textSubtle;
+  border-bottom: 0.063rem solid $colors-dark-cardBorder;
+  transform: translate3d(0px, 0px, 0px);
+  padding-left: 1rem;
+  padding-right: 1rem;
+
+  .nav-items {
+    display: flex;
+    height: 100%;
+
+    &-logo {
+      display: flex;
+      align-items: center;
+      height: 100%;
+
+      a {
+        display: flex;
+      }
+
+      img {
+        flex-shrink: 0;
+      }
+      // desktop's logo
+      img:first-of-type {
+        min-width: 10rem;
+        width: 10rem;
+        display: none;
+        @include lg {
+          display: block;
+        }
+      }
+      // mobile's logo
+      img:last-of-type {
+        min-width: 2rem;
+        width: 2rem;
+        @include lg {
+          display: none;
+        }
+      }
+    }
+
+    &-menu {
+      margin-left: 2.5rem;
+      height: 100%;
+      display: none;
+
+      @include md {
+        display: block;
+      }
+
+      ul {
+        display: flex;
+        align-items: center;
+        font-size: 1rem;
+        gap: 2rem;
+        height: 100%;
+
+        li > button {
+          @include aAsButton;
+        }
+
+        li:last-of-type button {
+          display: flex;
+          align-items: center;
+
+          img {
+            width: 1.25rem;
+          }
+
+          &::after {
+            @include lg {
+              content: '';
+              background: $colors-dark-success;
+              position: relative;
+              height: 0.5rem;
+              width: 0.5rem;
+              border-radius: 100%;
+              margin-left: 0.75rem;
+              display: block;
+            }
+          }
+        }
+      }
+    }
+  }
+
+  .nav-options {
+    height: 100%;
+    display: flex;
+    align-items: center;
+    font-weight: 600;
+
+    &-price {
+      display: none;
+      align-items: center;
+      margin-right: 0.75rem;
+
+      @include lg {
+        display: inline-flex;
+      }
+
+      .price-icon {
+        width: 1.5rem;
+        margin-right: 0.5rem;
+      }
+    }
+
+    &-language {
+      margin: 0.15rem 0.75rem 0 0;
+      padding: 0 0.5rem;
+    }
+
+    &-settings {
+      margin: 0.15rem 0.75rem 0 0;
+    }
+
+    &-chain {
+      display: inline-flex;
+      align-items: center;
+      background-color: $colors-dark-tertiary;
+      margin-right: 0.5rem;
+      color: $colors-dark-text;
+      border-radius: 16px;
+      box-shadow: $shadows-inset;
+      cursor: pointer;
+      height: 32px;
+      padding-left: 32px;
+      padding-right: 8px;
+      position: relative;
+
+      &-text {
+        margin-left: 0.5rem;
+        margin-right: 0.25rem;
+        display: none;
+
+        @include sm {
+          display: block;
+        }
+
+        .desktop-text {
+          display: none;
+
+          @include xl {
+            display: block;
+          }
+        }
+
+        .mobile-text {
+          @include xl {
+            display: none;
+          }
+        }
+      }
+
+      .bnb-chain-icon {
+        position: absolute;
+        left: 0;
+        top: 0;
+        width: 100%;
+        max-height: 2rem;
+        max-width: 2rem;
+      }
+
+      .down-arrow {
+        flex: 0;
+      }
+    }
+
+    &-wallet-connection {
+      height: 2rem;
+      display: inline-flex;
+      position: relative;
+      align-items: center;
+      justify-content: center;
+      padding: 0 1rem;
+      background-color: $colors-dark-primary;
+      color: $colors-dark-invertedContrast;
+      border: 0px;
+      border-radius: 1rem;
+      box-shadow: $shadows-inset;
+      cursor: pointer;
+      font-size: 1rem;
+      font-weight: 600;
+      letter-spacing: 0.03rem;
+
+      .desktop-text {
+        display: none;
+        line-height: 1;
+        @include md {
+          display: block;
+        }
+      }
+
+      .mobile-text {
+        @include md {
+          display: none;
+        }
+      }
+    }
+  }
+}
