+++ src/Assets/Components/Header/_Header.scss
@@ -1,0 +1,13 @@
+.head-area {
+  position: fixed;
+  top: 0px;
+  left: 0px;
+  transition: top 0.2s ease 0s;
+  height: 8.75rem; //140px
+  width: 100%;
+  z-index: 20;
+
+  @include sm {
+    height: 7.875rem; //120px
+  }
+}
