+++ src/Assets/Components/PhishingBanner/_PhishingBanner.scss
@@ -1,0 +1,121 @@
+.phishing-warning-container {
+  height: 5.25rem; //84px
+  min-height: 5.25rem;
+  max-height: 5.25rem;
+  width: 100%;
+  display: none;
+
+  @include sm {
+    height: 4.375rem; //70px
+    min-height: 4.375rem;
+    max-height: 4.375rem;
+  }
+
+  .phishing-banner {
+    overflow: hidden;
+    display: flex;
+    height: 100%;
+    padding: 0.75rem;
+    align-items: center;
+    background: $colors-dark-gradientPhishingDarker;
+
+    @include md {
+      background: $colors-dark-gradientPhishing;
+      padding: 0;
+    }
+
+    &-text-area {
+      @include md {
+        width: 100%;
+        height: 100%;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+      }
+    }
+
+    &-img {
+      display: none;
+
+      @include md {
+        display: block;
+        width: 5.75rem;
+        max-width: 100%;
+        height: auto;
+      }
+    }
+
+    &-close {
+      width: 1.5rem;
+      position: relative;
+      align-items: center;
+      border: 0rem;
+      border-radius: 1rem;
+      cursor: pointer;
+      display: inline-flex;
+      font-family: inherit;
+      font-size: 1rem;
+      font-weight: 600;
+      justify-content: center;
+      letter-spacing: 0.03em;
+      line-height: 1;
+      opacity: 1;
+      outline: 0rem;
+      transition: background-color 0.2s ease 0s, opacity 0.2s ease 0s;
+      height: 1.5rem;
+      padding: 0rem 1.1rem;
+      background-color: transparent;
+      color: $colors-dark-primary;
+      box-shadow: none;
+
+      @include md {
+        padding: 0rem 1.46rem;
+      }
+
+      img {
+        min-width: 1.25rem;
+        align-self: center;
+        fill: $colors-dark-white;
+        flex-shrink: 0;
+      }
+    }
+
+    &-text {
+      @include md {
+        width: 60%;
+        height: 80%;
+        display: flex;
+        align-items: center;
+        flex-wrap: wrap;
+        background: #27262c66;
+        border-radius: 1rem;
+        padding: 0.5rem;
+      }
+
+      .warning {
+        flex-shrink: 0;
+        color: $colors-light-textDisabled;
+        font-weight: 400;
+        line-height: 1.5;
+        font-size: 0.875rem;
+
+        @include md {
+          margin-right: 0.25rem;
+        }
+
+        &-important {
+          color: $colors-light-warning;
+          font-weight: 600;
+          line-height: 1.5;
+          text-transform: uppercase;
+          font-size: 0.875rem;
+        }
+        &-url {
+          color: $colors-light-white;
+          font-weight: 600;
+          line-height: 1.5;
+        }
+      }
+    }
+  }
+}
